adds 7 pkg extensions
