+++ Project.toml
@@ -4,22 +4,33 @@ authors = ["Daniel Rizk <rizk.daniel.12@gmail.com> and contributors"]
 version = "0.2.4"
 
 [deps]
-AWS = "fbe9abb3-538b-5e4e-ba9e-bc94f4f92ebc"
 Arrow = "69666777-d1a9-59fb-9406-91d4454c9d45"
 Chain = "8be319e6-bccf-4806-a6f7-6fae938471bc"
-ClickHouse = "82f2e89e-b495-11e9-1d9d-fb40d7cf2130"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DuckDB = "d2f5444f-75bc-4fdf-ac35-56f514c445e1"
 GZip = "92fee26a-97fe-5a0c-ad85-20a5f3185b63"
-GoogleCloud = "55e21f81-8b0a-565e-b5ad-6816892a5ee7"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
-LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
 ODBC = "be6f12e9-ca4f-5eb2-a339-a4f995cc0291"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+
+[weakdeps]
 SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
+LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
+GoogleCloud = "55e21f81-8b0a-565e-b5ad-6816892a5ee7"
+AWS = "fbe9abb3-538b-5e4e-ba9e-bc94f4f92ebc"
+MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
+ClickHouse = "82f2e89e-b495-11e9-1d9d-fb40d7cf2130"
+
+
+[extensions]
+SQLiteExt = "SQLite"
+LibPQExt = "LibPQ"
+GBQExt = "GoogleCloud"
+AWSExt = "AWS"
+MySQLExt = "MySQL"
+CHExt = "ClickHouse"
 
 [compat]
 AWS = "1.9"
@@ -44,6 +55,13 @@ julia = "1.9"
 [extras]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
+LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
+GoogleCloud = "55e21f81-8b0a-565e-b5ad-6816892a5ee7"
+AWS = "fbe9abb3-538b-5e4e-ba9e-bc94f4f92ebc"
+MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
+ClickHouse = "82f2e89e-b495-11e9-1d9d-fb40d7cf2130"
+
 
 [targets]
 test = ["Documenter", "Test"]
