+++ README.md
@@ -24,7 +24,7 @@ The main goal of TidierDB.jl is to bring the syntax of Tidier.jl to multiple SQL
 - Snowflake `set_sql_mode(:snowflake)`
 - Google Big Query `set_sql_mode(:gbq)`
 - Oracle `set_sql_mode(:oracle)`
-- Databricks
+- Databricks `set_sql_mode(:databricks)`
 
 The style of SQL that is generated can be modified using `set_sql_mode()`.
