+++ docs/examples/UserGuide/getting_started.jl
@@ -9,7 +9,7 @@
 
 # Alternatively, `using Tidier` will import TidierDB in the above manner for you, where TidierDB functions and macros will be available as `DB.@mutate()` and so on, and the TidierData equivalent would be `@mutate()`.
 
-# There are two ways to connect to the database. You can use `connect` without any need to load any additional packages. However, Oracle and Athena do not support this method yet and will require you to load in ODBC.jl or AWS.jl respectively.
+# To connect to a database, you can uset the `connect` function  as shown below, or establish your own connection through the respecitve libraries.
 
 # For example
 # Connecting to MySQL
@@ -21,14 +21,15 @@
 # conn = connect(:duckdb)
 # ```
 
-# Alternatively, you can use the packages outlined below to establish a connection through their respective methods.
+# ## Package Extensions 
+# The following backends utilize package extensions. To use one of backends listed below, you will need to write `using Library`
 
-# - ClickHouse: ClickHouse.jl
-# - MySQL and MariaDB: MySQL.jl
-# - MSSQL: ODBC.jl 
-# - Postgres: LibPQ.jl
-# - SQLite: SQLite.jl
-# - Athena: AWS.jl
-# - Oracle: ODBC.jl 
+# - ClickHouse: `using ClickHouse`
+# - MySQL and MariaDB: `using MySQL`
+# - MSSQL: `using ODBC` 
+# - Postgres: `using LibPQ``
+# - SQLite: `using SQLite`
+# - Athena: `using AWS`
+# - Oracle: `using ODBC` 
+# - Google BigQuery: `using GoogleCloud`
 
-# For DuckDB, SQLite, and MySQL, `copy_to()` lets you copy data to the database and query there. ClickHouse, MSSQL, and Postgres support for `copy_to()` has not been added yet.
